Moves Macro class into model package
